chore: remove most of the subgraph-compat code
