--- packages/ponder/src/subgraph-compat/runGraphCodegen.ts
@@ -1,10 +1,0 @@
-import { exec as execCallback } from "node:child_process";
-import util from "node:util";
-
-const exec = util.promisify(execCallback);
-
-const runGraphCodegen = async () => {
-  const { stdout, stderr } = await exec("graph codegen");
-};
-
-export { runGraphCodegen };
